refactor(frontend): is-link.jsと、同一の振る舞いをする記述を共通化
