+++ packages/frontend-embed/src/pages/clip.vue
@@ -51,7 +51,7 @@ import { misskeyApi } from '@/misskey-api.js';
 import { i18n } from '@/i18n.js';
 import { serverMetadata } from '@/server-metadata.js';
 import { url, instanceName } from '@@/js/config.js';
-import { isLink } from '@/to-be-shared/is-link.js';
+import { isLink } from '@@/js/is-link.js';
 import { defaultEmbedParams } from '@@/js/embed-page.js';
 import { DI } from '@/di.js';
