+++ packages/frontend-embed/src/pages/tag.vue
@@ -47,7 +47,7 @@ import EmTimelineContainer from '@/components/EmTimelineContainer.vue';
 import { i18n } from '@/i18n.js';
 import { serverMetadata } from '@/server-metadata.js';
 import { url, instanceName } from '@@/js/config.js';
-import { isLink } from '@/to-be-shared/is-link.js';
+import { isLink } from '@@/js/is-link.js';
 import { DI } from '@/di.js';
 import { defaultEmbedParams } from '@@/js/embed-page.js';
