+++ packages/frontend-shared/js/is-link.ts
